requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
python-multipart
aiofiles